refactor the metadata items to be in its own classes.
