+++ app/assets/javascripts/components/metadata/MetadataItem.jsx
@@ -1,0 +1,45 @@
+//app/assets/javascripts/components/MetadataItem.jsx
+var React = require('react');
+
+var MetadataString = require('./MetadataString');
+var MetadataDate = require('./MetadataDate');
+var MetadataHTML = require('./MetadataHTML');
+var MetadataText = require('./MetadataText');
+
+var fieldTypeMap = {
+  MetadataString: MetadataString,
+  MetadataDate: MetadataDate,
+  MetadataHTML: MetadataHTML,
+  MetadataText: MetadataText,
+};
+
+var MetadataItem = React.createClass({
+  displayName: 'Metadata Item',
+
+  propTypes: {
+    metadata: React.PropTypes.object.isRequired,
+  },
+
+  value: function(metadata_field) {
+    var MetadataComponent = fieldTypeMap[metadata_field["@type"]];
+    return (<MetadataComponent label={this.props.metadata.label} metadata_field={metadata_field} />);
+  },
+
+  map_arrays_to_values: function () {
+    return this.props.metadata.values.map(function (metadata_field) {
+      return this.value(metadata_field);
+    }, this);
+  },
+
+  render: function() {
+    return (
+      <dl>
+        <dt>{this.props.metadata.label}</dt>
+        <dd>{this.map_arrays_to_values()}</dd>
+      </dl>
+    );
+  }
+});
+
+// each file will export exactly one component
+module.exports = MetadataItem;
