+++ app/assets/javascripts/components/metadata/MetadataList.jsx
